Criando service e integração com a view (login)

+++ src/main/java/br/com/ies/candyShop/repository/EmpregadoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.ies.candyShop.entity.Empregado;
 
-public interface EmpregadoRepository extends JpaRepository<Empregado, Integer>{
+public interface EmpregadoRepository extends JpaRepository<Empregado, Long>{
 
 }
