+++ src/main/java/br/com/ies/candyShop/repository/EntregaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.ies.candyShop.entity.Entrega;
 
-public interface EntregaRepository extends JpaRepository<Entrega, Integer> {
+public interface EntregaRepository extends JpaRepository<Entrega, Long> {
 
 }
