+++ src/main/java/br/com/ies/candyShop/repository/PedidoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.ies.candyShop.entity.Pedido;
 
-public interface PedidoRepository extends JpaRepository<Pedido, Integer>{
+public interface PedidoRepository extends JpaRepository<Pedido, Long>{
 
 }
