+++ src/main/java/br/com/ies/candyShop/repository/ProdutoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.ies.candyShop.entity.Produto;
 
-public interface ProdutoRepository extends JpaRepository<Produto, Integer>{
+public interface ProdutoRepository extends JpaRepository<Produto, Long>{
 
 }
